% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\name{bartlett_eta}
\alias{bartlett_eta}
\title{Bartlett estimator}
\usage{
bartlett_eta(X)
}
\arguments{
\item{X}{A vector of estimates for the latent process. For example \eqn{\tilde{X}}.}
}
\description{
Bartlett estimator
}
