% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum.R
\name{changepoint_detection}
\alias{changepoint_detection}
\title{Detect changepoints using CUSUM test statistics}
\source{
Section 3 of \url{https://doi.org/10.1002/env.2752}
}
\usage{
changepoint_detection(Y, L, K = max(Y), t = 2:(length(Y) - 1), verbose = TRUE)
}
\arguments{
\item{Y}{Integer vector of responses.}

\item{L}{Period of the sinusoidal component of the latent process.}

\item{K}{Number of ordinal categories. \code{Y} is assumed to take on values \eqn{1, \dots, K}.}

\item{t}{Candidate positions for the changepoints. Defaults to \eqn{2, \dots, n - 1}.}
}
\value{
A vector containing the position of the best changepoint (\eqn{\tau}) and the respective value of the test statistic (\eqn{M}).
}
\description{
This function calculates a CUSUM test statistics at various time points and returns the highest one found.
This value can then be compared with (empirical) percentiles of the distribution of the test statistic,
to decide whether or not a changepoint has been found.
}
