% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_estimation}
\alias{model_estimation}
\title{Estimate the AOP(1) model from Li and Lu (2022)}
\source{
\url{https://doi.org/10.1002/env.2752}
}
\usage{
model_estimation(Y, K = max(Y))
}
\arguments{
\item{Y}{Integer vector.}

\item{K}{Integer number of response categories. If not provided, assumed to be \code{max(Y)}.}
}
\description{
The parameter vector is \eqn{\theta = (c_2, \dots, c_{K-1}, \alpha_0, \alpha_1, A, D, \Delta)}.
Starting values for \eqn{\alpha_1, A, D, \Delta} are set to 0.
Starting values for \eqn{\alpha_0} and thresholds \eqn{c_2,\dots,c_{K-1}} are \eqn{\alpha_0 = -\Phi^{-1}(p_1)} and \eqn{c_k=-\Phi^{-1}\left(\sum_{j=1}^k p_j\right) + \alpha_0}, where \eqn{p_j} is the sample proportion of times that \eqn{Y_t} lies in category \eqn{j}.
}
