% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{ordinalchange_stan}
\alias{ordinalchange_stan}
\title{Fit an autoregressive ordinal probit model using Stan}
\usage{
ordinalchange_stan(Y, K, period, ...)
}
\arguments{
\item{Y}{Integer vector of ordered categorical responses.}

\item{K}{Total number of response categories (assumed to go \eqn{1,\dots,K}).}

\item{period}{The period of the seasonal component of the latent process.}

\item{...}{Further arguments passed to \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}}.}
}
\description{
This is an alternative to the sequential estimation procedure proposed by Li and Mo.
See \url{https://mc-stan.org/docs/stan-users-guide/change-point.html} to better understand the implementation.
}
\details{
Internally we are using \code{optimizing}, i.e. fitting the model using L-BFGS optimization, rather than \code{sampling},
which performs Hamiltonian Monte Carlo (automatic differentation). This is because the model is otherwise extremely slow to fit.
However, it may be possible to optimize the Stan code such that MCMC becomes more feasible in future.
}
