% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Q}
\alias{Q}
\title{Sum of squared discrepancies between observations and their expected values}
\usage{
Q(theta, Y, tau, L)
}
\arguments{
\item{theta}{A vector of parameters.}

\item{Y}{A vector of ordered categorical responses.}

\item{tau}{A candidate changepoint.}

\item{L}{The known period of the seasonal component.}
}
\description{
Sum of squared discrepancies between observations and their expected values
}
\note{
The symbol \code{T} has been replaced with \code{L} to avoid confusion with \code{T == TRUE}.
}
