% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{seq_estimation}
\alias{seq_estimation}
\title{Estimate the AOP(1) model from Li and Lu (2022)}
\source{
\url{https://doi.org/10.1002/env.2752}
}
\usage{
seq_estimation(Y, tau, K = max(Y), L = 7, ...)
}
\arguments{
\item{Y}{Integer vector.}

\item{tau}{Time of the proposed changepoints.}

\item{K}{Integer number of response categories, denoted \eqn{k = 1, \dots, K}.}

\item{L}{Period of the sinusoidal component.}

\item{...}{Other arguments, ignored.}
}
\description{
We use the sequential estimation approach described in that paper.
The latent autocorrelation parameter \eqn{\rho} is estimated separately from the other parameters.
}
\details{
The parameter vector is \eqn{\theta = (c_2, \dots, c_{K-1}, \alpha_0, \alpha_1, A, D, \Delta)}.
Starting values for \eqn{\alpha_1, A, D, \Delta} are set to 0.
Starting values for \eqn{\alpha_0} and thresholds \eqn{c_2,\dots,c_{K-1}} are \eqn{\alpha_0 = -\Phi^s that \eqn{Y_t} lies in category \eqn{j}.
{-1}(p_1)} and \eqn{c_k=-\Phi^{-1}(\sum_{j=1}^k p_j) + \alpha_0}, where \eqn{p_j} is the sample proportion of time
}
\examples{
Y <- sample(1:3, 50, replace = T)
seq_estimation(Y)

}
